chore: move all vega subcommand under cmd/vega/commands
