+++ cmd/vega/commands/bridge/privkey_signer.go
