+++ cmd/vega/commands/genesis/generate.go
