+++ cmd/vega/commands/genesis/new_validator.go
