+++ cmd/vega/commands/node/app_wrapper.go
