+++ cmd/vega/commands/paths/list.go
