+++ cmd/vega/commands/query.go
@@ -10,12 +10,12 @@
 // of this software will be governed by version 3 or later of the GNU General
 // Public License.
 
-package main
+package commands
 
 import (
 	"context"
 
-	"code.vegaprotocol.io/vega/cmd/vega/query"
+	"code.vegaprotocol.io/vega/cmd/vega/commands/query"
 	"github.com/jessevdk/go-flags"
 )
