+++ cmd/vega/commands/query/assets.go
