+++ cmd/vega/commands/query/markets.go
