+++ cmd/vega/commands/query/validators.go
