+++ cmd/vega/commands/subcmd_example.go
