+++ cmd/vega/commands/unsafe_reset_all.go
@@ -10,7 +10,7 @@
 // of this software will be governed by version 3 or later of the GNU General
 // Public License.
 
-package main
+package commands
 
 import (
 	"context"
@@ -28,7 +28,6 @@ type UnsafeResetAllCmd struct {
 	config.VegaHomeFlag
 }
 
-//nolint:unparam
 func (cmd *UnsafeResetAllCmd) Execute(_ []string) error {
 	log := logging.NewLoggerFromConfig(
 		logging.NewDefaultConfig(),
