+++ cmd/vega/commands/verify.go
@@ -10,12 +10,12 @@
 // of this software will be governed by version 3 or later of the GNU General
 // Public License.
 
-package main
+package commands
 
 import (
 	"context"
 
-	"code.vegaprotocol.io/vega/cmd/vega/verify"
+	"code.vegaprotocol.io/vega/cmd/vega/commands/verify"
 
 	"github.com/jessevdk/go-flags"
 )
