+++ cmd/vega/commands/verify/asset.go
