+++ cmd/vega/commands/verify/genesis_test.go
