+++ cmd/vega/commands/verify/reporter.go
