+++ cmd/vega/commands/verify/verifier.go
