+++ cmd/vega/main.go
@@ -14,87 +14,14 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"os"
 
-	"github.com/jessevdk/go-flags"
-
-	"code.vegaprotocol.io/vega/cmd/vega/faucet"
-	"code.vegaprotocol.io/vega/cmd/vega/genesis"
-	"code.vegaprotocol.io/vega/cmd/vega/nodewallet"
-	"code.vegaprotocol.io/vega/cmd/vega/paths"
-	"code.vegaprotocol.io/vega/core/config"
+	"code.vegaprotocol.io/vega/cmd/vega/commands"
 )
-
-// Subcommand is the signature of a sub command that can be registered.
-type Subcommand func(context.Context, *flags.Parser) error
-
-// Register registers one or more subcommands.
-func Register(ctx context.Context, parser *flags.Parser, cmds ...Subcommand) error {
-	for _, fn := range cmds {
-		if err := fn(ctx, parser); err != nil {
-			return err
-		}
-	}
-	return nil
-}
 
 func main() {
 	ctx := context.Background()
-	if err := Main(ctx); err != nil {
+	if err := commands.Main(ctx); err != nil {
 		os.Exit(-1)
 	}
-}
-
-func Main(ctx context.Context) error {
-	// special case for the tendermint subcommand, so we bypass the command line
-	if len(os.Args) >= 2 {
-		switch os.Args[1] {
-		case "tendermint", "tm":
-			return (&tmCmd{}).Execute(nil)
-		case "wallet":
-			return (&walletCmd{}).Execute(nil)
-		case "datanode":
-			return (&datanodeCmd{}).Execute(nil)
-		case "blockexplorer":
-			return (&blockExplorerCmd{}).Execute(nil)
-		case "tools":
-			return (&toolsCmd{}).Execute(nil)
-		}
-	}
-
-	parser := flags.NewParser(&config.Empty{}, flags.Default)
-
-	if err := Register(ctx, parser,
-		faucet.Faucet,
-		genesis.Genesis,
-		Init,
-		nodewallet.NodeWallet,
-		Verify,
-		Version,
-		Wallet,
-		Datanode,
-		VegaTools,
-		Watch,
-		Tm,
-		Tendermint,
-		Query,
-		Bridge,
-		paths.Paths,
-		UnsafeResetAll,
-		AnnounceNode,
-		RotateEthKey,
-		ProposeProtocolUpgrade,
-		Start,
-		Node,
-		BlockExplorer,
-	); err != nil {
-		_, _ = fmt.Fprintln(os.Stderr, err)
-		return err
-	}
-
-	if _, err := parser.Parse(); err != nil {
-		return err
-	}
-	return nil
 }
